editoast: rearchitecture editoast_authz
* Moves `Regulator` and `StorageDriver` in their own module
* Moves user and group related structs into a `subject` module
* Move the `Error<T>` type in `lib.rs`
* Re-exposes more definitions at crate-level and makes some sub-modules
  private
