+++ editoast/editoast_authz/src/subject.rs
@@ -1,0 +1,26 @@
+pub type UserIdentity = String;
+pub type UserName = String;
+pub type GroupName = String;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct UserInfo {
+    pub identity: UserIdentity,
+    pub name: UserName,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct GroupInfo {
+    pub name: GroupName,
+}
+
+impl std::fmt::Display for UserInfo {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} ({})", self.identity, self.name)
+    }
+}
+
+impl std::fmt::Display for GroupInfo {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.name)
+    }
+}
