+++ editoast/src/client/roles.rs
@@ -6,10 +6,10 @@ use anyhow::anyhow;
 use anyhow::bail;
 use clap::Args;
 use clap::Subcommand;
-use editoast_authz::authorizer::GroupInfo;
-use editoast_authz::authorizer::StorageDriver;
-use editoast_authz::authorizer::UserInfo;
+use editoast_authz::subject::GroupInfo;
+use editoast_authz::subject::UserInfo;
 use editoast_authz::BuiltinRole;
+use editoast_authz::StorageDriver;
 use editoast_models::DbConnectionPoolV2;
 use itertools::Itertools as _;
 use strum::IntoEnumIterator;
